+++ gEconpy/numba_tools/intrinsics.py
